Implement "Rename account"
