+++ app/src/main/kotlin/at/bitfire/davdroid/ui/account/AccountModel.kt
@@ -1,14 +1,20 @@
 package at.bitfire.davdroid.ui.account
 
+import android.Manifest
 import android.accounts.Account
 import android.accounts.AccountManager
 import android.accounts.OnAccountsUpdateListener
+import android.annotation.SuppressLint
 import android.app.Application
+import android.content.ContentResolver
 import android.content.pm.PackageManager
 import android.os.Handler
 import android.os.Looper
 import android.provider.CalendarContract
 import android.provider.ContactsContract
+import androidx.annotation.MainThread
+import androidx.annotation.WorkerThread
+import androidx.core.content.ContextCompat
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
@@ -20,20 +26,26 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.work.WorkInfo
 import at.bitfire.dav4jvm.DavResource
+import at.bitfire.davdroid.InvalidAccountException
 import at.bitfire.davdroid.R
 import at.bitfire.davdroid.db.AppDatabase
 import at.bitfire.davdroid.db.Collection
 import at.bitfire.davdroid.db.Service
 import at.bitfire.davdroid.log.Logger
 import at.bitfire.davdroid.network.HttpClient
+import at.bitfire.davdroid.resource.LocalAddressBook
+import at.bitfire.davdroid.resource.LocalTaskList
 import at.bitfire.davdroid.resource.TaskUtils
 import at.bitfire.davdroid.servicedetection.RefreshCollectionsWorker
 import at.bitfire.davdroid.settings.AccountSettings
+import at.bitfire.davdroid.syncadapter.AccountsCleanupWorker
+import at.bitfire.davdroid.syncadapter.PeriodicSyncWorker
 import at.bitfire.davdroid.syncadapter.SyncWorker
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.launch
 import java.util.Optional
 import java.util.logging.Level
@@ -116,6 +128,7 @@ class AccountModel @AssistedInject constructor(
     val calendarsPager = CollectionPager(db, calDavSvc, Collection.TYPE_CALENDAR, showOnlyPersonal)
     val webcalPager = CollectionPager(db, calDavSvc, Collection.TYPE_WEBCAL, showOnlyPersonal)
 
+    val renameAccountError = MutableLiveData<String>()
     val deleteCollectionResult = MutableLiveData<Optional<Exception>>()
 
 
@@ -135,6 +148,145 @@ class AccountModel @AssistedInject constructor(
 
 
     // actions
+
+    /**
+     * Will try to rename the [account] to given name.
+     *
+     * @param newName new account name
+     */
+    fun renameAccount(newName: String) {
+        val oldAccount = account
+
+        // remember sync intervals
+        val oldSettings = try {
+            AccountSettings(context, oldAccount)
+        } catch (e: InvalidAccountException) {
+            renameAccountError.postValue(context.getString(R.string.account_invalid))
+            invalid.postValue(true)
+            return
+        }
+
+        val authorities = mutableListOf(
+            context.getString(R.string.address_books_authority),
+            CalendarContract.AUTHORITY
+        )
+        TaskUtils.currentProvider(context)?.authority?.let { authorities.add(it) }
+        val syncIntervals = authorities.map { Pair(it, oldSettings.getSyncInterval(it)) }
+
+        val accountManager = AccountManager.get(context)
+        // check whether name is already taken
+        if (accountManager.getAccountsByType(context.getString(R.string.account_type)).map { it.name }.contains(newName)) {
+            Logger.log.log(Level.WARNING, "Account with name \"$newName\" already exists")
+            renameAccountError.postValue(context.getString(R.string.account_rename_exists_already))
+            return
+        }
+
+        try {
+            /* https://github.com/bitfireAT/davx5/issues/135
+            Lock accounts cleanup so that the AccountsCleanupWorker doesn't run while we rename the account
+            because this can cause problems when:
+            1. The account is renamed.
+            2. The AccountsCleanupWorker is called BEFORE the services table is updated.
+               → AccountsCleanupWorker removes the "orphaned" services because they belong to the old account which doesn't exist anymore
+            3. Now the services would be renamed, but they're not here anymore. */
+            AccountsCleanupWorker.lockAccountsCleanup()
+
+            // Renaming account
+            accountManager.renameAccount(oldAccount, newName, @MainThread {
+                if (it.result?.name == newName /* account has new name -> success */)
+                    viewModelScope.launch(Dispatchers.Default + NonCancellable) {
+                        try {
+                            onAccountRenamed(accountManager, oldAccount, newName, syncIntervals)
+                        } finally {
+                            // release AccountsCleanupWorker mutex at the end of this async coroutine
+                            AccountsCleanupWorker.unlockAccountsCleanup()
+                        }
+                    } else
+                // release AccountsCleanupWorker mutex now
+                    AccountsCleanupWorker.unlockAccountsCleanup()
+
+                // close AccountActivity with old name
+                invalid.postValue(true)
+            }, null)
+        } catch (e: Exception) {
+            Logger.log.log(Level.WARNING, "Couldn't rename account", e)
+            renameAccountError.postValue(context.getString(R.string.account_rename_couldnt_rename))
+        }
+    }
+
+    /**
+     * Called when an account has been renamed.
+     *
+     * @param oldAccount the old account
+     * @param newName the new account
+     * @param syncIntervals map with entries of type (authority -> sync interval) of the old account
+     */
+    @SuppressLint("Recycle")
+    @WorkerThread
+    fun onAccountRenamed(accountManager: AccountManager, oldAccount: Account, newName: String, syncIntervals: List<Pair<String, Long?>>) {
+        // account has now been renamed
+        Logger.log.info("Updating account name references")
+        val context: Application = getApplication()
+
+        // disable periodic workers of old account
+        syncIntervals.forEach { (authority, _) ->
+            PeriodicSyncWorker.disable(context, oldAccount, authority)
+        }
+
+        // cancel maybe running synchronization
+        SyncWorker.cancelSync(context, oldAccount)
+        for (addrBookAccount in accountManager.getAccountsByType(context.getString(R.string.account_type_address_book)))
+            SyncWorker.cancelSync(context, addrBookAccount)
+
+        // update account name references in database
+        try {
+            db.serviceDao().renameAccount(oldAccount.name, newName)
+        } catch (e: Exception) {
+            Logger.log.log(Level.SEVERE, "Couldn't update service DB", e)
+            renameAccountError.postValue(context.getString(R.string.account_rename_couldnt_rename))
+            return
+        }
+
+        // update main account of address book accounts
+        if (ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_CONTACTS) == PackageManager.PERMISSION_GRANTED)
+            try {
+                context.contentResolver.acquireContentProviderClient(ContactsContract.AUTHORITY)?.use { provider ->
+                    for (addrBookAccount in accountManager.getAccountsByType(context.getString(R.string.account_type_address_book))) {
+                        val addressBook = LocalAddressBook(context, addrBookAccount, provider)
+                        if (oldAccount == addressBook.mainAccount)
+                            addressBook.mainAccount = Account(newName, oldAccount.type)
+                    }
+                }
+            } catch (e: Exception) {
+                Logger.log.log(Level.SEVERE, "Couldn't update address book accounts", e)
+            }
+
+        // calendar provider doesn't allow changing account_name of Events
+        // (all events will have to be downloaded again)
+
+        // update account_name of local tasks
+        try {
+            LocalTaskList.onRenameAccount(context, oldAccount.name, newName)
+        } catch (e: Exception) {
+            Logger.log.log(Level.SEVERE, "Couldn't propagate new account name to tasks provider", e)
+        }
+
+        // retain sync intervals
+        val newAccount = Account(newName, oldAccount.type)
+        val newSettings = AccountSettings(context, newAccount)
+        for ((authority, interval) in syncIntervals) {
+            if (interval == null)
+                ContentResolver.setIsSyncable(newAccount, authority, 0)
+            else {
+                ContentResolver.setIsSyncable(newAccount, authority, 1)
+                newSettings.setSyncInterval(authority, interval)
+            }
+        }
+
+        // synchronize again
+        SyncWorker.enqueueAllAuthorities(context, newAccount)
+    }
+
 
     /** Deletes the account from the system (won't touch collections on the server). */
     fun deleteAccount() {
